+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/IdentifierListInfo.kt
@@ -13,18 +13,9 @@ data class IdentifierListInfo(
     val identifier: ByteArray,
     @SerialName("uri")
     override val uri: UniformResourceIdentifier,
-    /**
-     * The identifier_list and status_list in the MSO may contain the Certificate element. If the
-     * Certificate element is present, it shall contain a certificate containing the public key that signed the
-     * top-level certificate in the x5chain element in the MSO revocation list structure. The mdoc reader shall
-     * use that certificate as trust point for verification of the x5chain element in the MSO revocation list
-     * structure. If the Certificate element is not present, the top-level certificate in the x5chain element
-     * shall be signed by the certificate used to sign the certificate in the x5chain element of the MSO. In the
-     * context of an mDL, that is the IACA certificate.
-     */
     @SerialName("certificate")
     @ByteString
-    val certificate: ByteArray? = null,
+    override val certificate: ByteArray? = null,
 ) : RevocationListInfo() {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true